recordjar: decode.KeyedStringList now returns map[string]string
The decode.KeyedStringList function currently returns a slice of string
arrays each two elements in length. This leads to the elements being
accessed using 'magic' numbers such as pair[0] and pair[1] when working
with the returned values. Accessing the pairs also requires looping over
the slice to retrieve the arrays, and then accessing the array elements.

This change replaces [][2]string with a map[string]string which is
simpler to range over and requires no 'magic' numbers in the code.

The only change in functionality is that the map cannot have duplicate
keys. This means that a duplicate name on the left of a pair will
overwrite a previous value:

  Vetoes: GET→You go to get it, but it gets up and runs away.
        : GET→You cannot get that.

In this example "GET" would have the value "You cannot get that". As
duplicate names are not currently handled or expected this is not an
issue.

This change also updates the vetoes attribute to use the new returned
type.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>