recordjar: decode.KeyedString now returns name & value strings
The decode.KeyedString function currently returns an array of two
strings for the name and value pair decoded. This leads to the elements
being accessed using 'magic' numbers such as pair[0] and pair[1] when
working with the returned values.

This change replaces [2]string with a name string and a value string
which is simpler and requires no 'magic' numbers in the code.

This change also updates KeyedStringList to use the new returned types.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>