recordjar: encode.PairList now takes map[string]string & rune delimeter
The encode.PairList function currently takes a slice of string arrays
each two elements in length. This leads to the elements being accessed
using 'magic' numbers such as pair[0] and pair[1] when working with the
passed values. Accessing the pairs also requires looping over the slice
to retrieve the arrays, and then accessing the array elements.

This change replaces [][2]string with a map[string]string which is
simpler to range over and requires no 'magic' numbers in the code.

The delimeter has also been updated from a string to a rune as we only
accept single character delimeters.

There is no change in functionality as this function is currently
unused.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>