recordjar: Normalise Record field names in Jar.Write
When comparing field names there is sometimes a discrepancy in the field
name given in the Record map and what is used in comparisons. This is
especially true when it comes to the name of the freetext field.

This change populates a map of normalised field names with the key as
the name as presented and the value the normalised name. Comparisons are
then made by looking up the normalised name.

A normalised field name is one that has been lowercased and then the
first letter title cased.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>