attr,has: Add has/thing.Collectable and implement it
This change adds a Collectable method to the has.Thing interface and
implements it in attr.Thing. The Collectable method returns true if the
Thing is collectable by players and can be saved in their player file.
If false is returned the Thing can only be possessed by a player
temporally and will be 'lost' when the player quits if they still have
the item in their inventory.

The dump method of a Thing now reports whether it is collectable or not.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>