cmd: Add SAVE command (WIP)
This command adds a SAVE command to save a player and their inventory to
a .wrj file. At the moment the player file written out cannot be loaded
again as the account information record is missing as is the loading of
a player's inventory. For testing, and until all of the other pieces are
in place, the SAVE command will only write out to the file save.wrj in
the players directory.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>