cmd: use Thing.Collectable & on the fly resets in junk.dispose
Currently an item can be kept/collected by a player if it has no Reset
attribute and no origin. This is being changed so that an item can now
be kept/collected by a player if Thing.Collectable returns true. The
implication of this is that when junking a Thing, collectable items
should be disposed of permanently and garbage collected. Non-collectable
items should always be reset. However a Thing may not have a Reset
attribute associated with it.

If there is no Reset attribute available junk.dispose will now schedule
a $RESET event on the fly. This is safe as the Reset attribute only
controls the scheduling of a $RESET and it is $RESET that performs the
actual resetting. The delay between junking and reset is currently one
second - the minimum delay time imposed by event.Queue.

An alternative would be to create a Reset attribute on the fly and add
it to the Thing before calling Reset. Or to add a default Reset
attribute to any Thing that does not specify one. In both cases this
would cause an increase in memory usage. This could possibly allow the
removal of a few ugly corner cases.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>