cmd: QUIT now saves a player and removes items by junking
Currently the QUIT command will drop all of a player's Inventory items
when a player leaves the world.

This change will now save the player and all of their collectable items
to the player file. It will then junk all of the inventory items so that
they are either disposed of and garbage collected or reset.

It should be noted that player files are currently only written to the
file 'players/save.wrj' and can not currently be loaded when the player
logs back into the server. Work is ongoing to be able to load player
files complete with items. When this is ready the SAVE command will save
real player files and logging in will load them.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>