attr: Add account information to the Player attribute
In order to be able to write out a complete player file the account
information needs to be available in order to write the file's header
record. By holding the information in the player attribute, instead of
in the frontend, the SAVE command has access to all of the information
required in order to write the player file.

This change adds an account type with associated Marshal, Unmarshal and
Set methods for manipulating the information held.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>