frontend: Update account.write to use attributes and SAVE command
Currently the account.write method directly builds a Jar and handles
writing it out to disk. It then unmarshals the Jar to create the player.

This change builds a player by adding attributes to a Thing and saves it
using the SAVE command. This allows for the error prone manual Jar
building to be dropped. Most of the file and associated error handling
is also dropped.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>