cmd: Fix empty inventory when re-entering game after QUIT
When a QUIT command is issued all items are currently junked from the
player's inventory. However if they re-enter the game from the main menu
their inventory remains empty - the player's save file is only loaded
once, during login.

This change updates the QUIT command so that only non-collectable items
are disposed of, leaving the rest of the player's inventory intact. The
player's collectable items are already disposed of by the frontend.Close
method when the player either logs out or if there is a network error.

As only non-collectable items are now disposed of during QUIT, we can
notify the player of the items being junked. Non-collectable items can
also be disposed of before the SAVE command is processed - potentially
resulting in fewer items being scanned during the save.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>