frontend: Restore player inventory with Thing copies from temp store
This change fixes an benign issue whereby a Thing can be freed twice.

When loading a player save file a temporary store of Things is created.
The player's inventory should then be populated with copies of Things
from the temporary store. This allows duplicate references, where one
item definition can be referenced in multiple places, to be resolved as
separate item instances.

If the inventory is not rebuilt using copies of Things the Things will
be freed twice - once when the temporary store is cleared down and again
when the Thing is disposed of.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>