attr,frontend: Fix Thing count broken by temporary stores
This is the 3rd commit to fix Thing accounting. The previous fixes are:

  9f028c3: frontend: Restore player inventory with Thing copies from
  temp store

  d8a091e: frontend: login.assemblePlayer - enable copied, not store
  items

The fix for 9f028c3 was incorrect, the temporary store should be
populated using references and not copies. This is so that complex
containers with containers can be resolved when copies are made from the
temporary store.

Commit d8a091e perpetuated the incorrect use of copies instead of using
references in the temporary store.

The cause of the issue is that when a reference is used Thing.ThingCount
is only incremented once, when the original Thing is created. However
Thing.Free is called for each reference, decrementing Thing.ThingCount
multiple times. This causes the skew in the count of the number of
Things that exist.

Why is accounting for the number of Thing important? In general use only
a copy of a Thing should be put into the world. Monitoring the count of
Thing can give an indication that references are accidentally being used
or that a Thing is being freed when it should not.

This commit adds a Thing.freed flag that can be checked to see if a
Thing has already been freed. If it has the Thing.ThingCount is not
decremented. However this then masks real issues when a Thing is freed
multiple times when it should only be freed once - see previous
paragraph. It also adds the storage overhead of a bool flag for every
Thing. These two issues will be addressed in a another commit, for now
the focus is on correcting the current Thing count.

It would be useful to report on a Thing that is freed when it should not
be, indicating accidental use of a Thing reference in the main game
world.

This commit also, correctly, reverts to using references instead of
copies for the temporary store.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>