attr,has: Add Thing.NotUnique + update Thing.Free to use it
This change adds a new NotUnique method to Thing. The method provides a
way for a Thing to be marked as not being unique. This is so that,
possibly multiple, references to a Thing can be used when creating
temporary stores - such as when loading zones or players. This addresses
both of the concerns of commit b128c1e: It drops the additional storage
required for the freed flag, as it is no longer required. It also
provides a warning if a unique Thing is freed more than once.

If a Thing is marked as non-unique the ThingCount will be decremented by
one, to account for the increase when NewThing is called. Subsequent
calls to Free will not decrement the ThingCount. Calling NotUnique
multiple times on the same Thing is safe.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>