frontend,zones: Mark Thing in temporary stores non-unique
When building temporary stores for loading zones or players, mark Things
added as non-unique. This allow references to be used to resolve complex
container in container hierarchies while keeping an acurate thing count.
It also suppresses warnings about a Thing being freed multiple times, as
this is expected for non-unique Thing.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>