attr: Drop temp Marshal method from Attribute type
During development of the attributes it was convenient to add a Marshal
method to the Attribute type so that it didn't need to be implemented by
all attributes at once.

Now that most attributes have a Marshal method this change implements
Marshal for the remaining attributes: Locate and Player, and drops it on
Attribute. The compiler will now warn if an attribute does not implement
Marshaler.

Dropping Marshal on Attribute also stops the message "[DEBUG] dummy
marshal" from being logged.

NOTE: At the moment dummy methods have been added to Locate and Player
to implement the Marshaler interface. This is ugly and it should be
possible to remove them after some remaining work on the Marshaler is
completed.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>