comms: Log using sequence number not incoming IP address/port
This commit addresses concerns raised about logging the incoming IP
address and source port of players in relation to compliance with the
General Data Protection Regulation (GDPR) (EU) 2016/679.

Instead of logging messages identified by incoming IP address and source
port number the messages will now be identified by a sequence number
unique to each connection. This will allow a series of log entries to be
identified and followed so that a picture of what occurred during a
session can be built up when debugging:

  [1] connection from 127.0.0.1:43834
  [2] connection from 127.0.0.1:43836
  [1] connection error: frontend closed
  [1] connection closed
  [2] connection error: read tcp 127.0.0.1:4001->127.0.0.1:43836: i/o timeout
  [3] connection from 127.0.0.1:43838
  [3] connection dropped by remote client
  [2] connection closed
  [3] connection closed

Here we have three connections, identified by [1],[2] and [3]. A future
commit will add a configuration setting that will turn off the logging
of IP addresses and port numbers completely, dropping the 'connection
from' lines.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>