comms,config,data: Add Server.LogClient config knob
This commit addresses concerns raised about logging the incoming IP
address and source port of players in relation to compliance with the
General Data Protection Regulation (GDPR) (EU) 2016/679.

This change adds a new configuration setting Server.LogClient. When set
to false the IP address and source port number of a connecting client
will not be logged in the server logs. When set to true the IP address
and source port number will be logged. An example session:

  2018/06/16 17:04:46 [123] connection from 127.0.0.1:59692
  2018/06/16 17:06:31 [123] connection error: frontend closed
  2018/06/16 17:06:31 [123] connection closed

The IP address and source port are logged together with the sequence
number allocated to the connection. In this case the sequence number is
123. Without the IP address and source port it is still possible to see
what occurred during a client session by filtering the log by sequence
number, without identifying the incoming connection.

The configuration setting Server.LogClient defaults to false, to not log
the incoming IP address and source port.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>