comms: Don't leak IP addresses in errors if not logging IPs
Some error messages from the standard library contain the IP address and
ports of connecting clients:

  connection error: write tcp 127.0.0.1:4001->127.0.0.1:36200: \
  write: broken pipe

If we are not supposed to be logging client IP addresses then replace
the IP address and port with '???' when logging the message.

The documentation for Server.LogClient in configuration-file.txt has
been updated to reflect this change.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>