frontend: Switch frontend to per-connection logging
This change updates the frontend type to include a log.Conn
per-connection logger. The New function has been updated to accept a
log.Conn which is then saved in the new instance of the frontend type
created. The log.Conn can then be accessed in the frontend to log
per-connection messages.

Comment for New updated to include passed log.Conn and some rewording to
improve the comment overall.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>