recordjar: Handle continuation lines in StringList/KeyedStringList better
Currently string lists are simply concatenated together:

  OnAction: The cat curls up and starts to purr like a buzz-saw. : The
            cat starts to claw at the furniture, scratching deep gouges
            into the wood. : The cat starts to wash itself.

However, the colon separator should be on a new line and indented level
with the field/data separator:

  OnAction: The cat curls up and starts to purr like a buzz-saw.
          : The cat starts to claw at the furniture, scratching deep
            gouges into the wood.
          : The cat starts to wash itself.

This change fixes the formatting. It should be noted that there is one
nit, outdenting the colon separator frees up two character positions on
a line. However, the lines are not refolded.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>