recordjar: encode.PairList sort + trim keys & values
A pair list should be consistently sorted for reproducibility. Both the
key and value of a pair should have all white space removed - leading,
trailing or within. A pair with a key with white space would be read as
a key with no value followed by a key/value pair. A pair with a value
with white space would be read as a key/value pair followed by a key
with no value. A pair with a key and a value, both with white space
would be read as... a key with no value, followed by a key/value pair,
followed by a key with no value. None of these three outcomes would be
correct.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>