recordjar: Modify indexSeparator, Simplify KeyedString+PairList(+no dups)
Currently indexSeparator returns a position of -1 if no separator is
found. However by returning the full length of the passed slice instead
the returned values can be used directly for slicing without having to
check the return values. This leads to a simplification of the PairList
and KeyedString functions.

In addition PairList has been updated to ignore duplicate keywords after
the first instance.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>