recordjar: Stop decode.Bytes trimming white space at '\n'.
The decode.Bytes should trim white space but should not remove line
feeds '\n'. This allows for leading and trailing blank lines to be
included in the data. This is the preferred way to decode a free text
section - rather than using decode.String which trims all leading and
trailing white space including line feeds. Comment added explaining this
to decode.String.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>