recordjar: encode.PairList drop nameless value, fix empty value check
When encoding a PairList ignore values where no name is given. That is,
if the name - processed as a keyword, with white space removal - is an
empty string any value associated with it will be ignored.

Checking values to see if they are empty strings should be done after
trimming leading and trailing white space.

This change also drops use of a string.Builder in favour of a simple
[]byte which is now more efficient with the above changes.

Tests have been updated for new expectations. The check for an empty
name also resolves the issue of the follow test I was previously
uncertain about:

  {map[string]string{"": "z"}, '→', "→z"}, // correct?

The test is now:

  {map[string]string{"": "z"}, '→', ""},

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>