recordjar: encode/decode - Don't call KeyedString in KeyedStringList
There is no rationale for having a KeyedString function. There is no
situation in which you would want to use a KeyedString instead of a
KeyedStringList.

This change drops the use of KeyedString from within KeyedStringList in
the encoder and decoder. This will make it possible to drop KeyedString
from the encoder and decoder.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>