docs: drop mention of keyed strings, expand key string lists
There is no rationale for having a KeyedString function. There is no
situation in which you would want to use a KeyedString instead of a
KeyedStringList.

This change drops keyed strings from the documentation. It also expands
the description for keyed string lists, which previously referred to
keyed strings for part of its description.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>