recordjar: Separate+fix side effect tests for encode/decode Bytes
The TestBytes tests for encoding/decoding Bytes is being overly
complicated by the fact we are trying to test for side effects at the
same time as testing the results of encoding/decoding bytes.

This change turns TestBytes for encoding and decoding into very simple
tests. The tests for side effects is then moved into TestBytesSideEfects
for both encoding and decoding.

In TestBytesSideEfects the check to make sure a copy is returned is now
accomplished by overwriting the returned result and observing if the
original backing array is modified. This is due to the fact that finding
the maximum end of the slices using &x[0:cap(x)][cap(x)-1] and comparing
them is unreliable with the introduction of full index expressions -
which can modify the capacity of a slice, and hence they will not match
even if the slices both have the same backing array.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>