cmd: Add $ACT command for performing actions
The ACT commands provides a way to act something out. For example, "act
produces a red and white spotted hankie." would show the message:

  Diddymus produces a red and white spotted hankie.

The same message is shown to the actor and any observers. If the name of
the actor starts with "A " it is changes to "The ". For example if we
have a name of "a clock" and it issues "$ACT chimes", then the message
would be "The clock chimes." and not "A clock chimes."

This is currently a scripting only command as its introduction for
players, in the Java version, was rather... contentious.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>