attr,cmd,docs: Change OnAction string lists to be scriptable commands
Actions are currently a simple list of strings defined using OnAction.
When an action fires, as defined by Action, a random string is picked
from the list and displayed.

This change modifies OnAction so that they are now treated as a list of
commands that are scripted. This allows for a much wider range of simple
actions. It also allows for simple mobiles to be created that can randomly
wander around - by scripting 'N', 'E' etc. For example:

	 Ref: M3
      Name: a small frog
     Alias: FROG
     Reset: AFTER→1m
   OnReset: A frog hops into view.
    Action: AFTER→10s JITTER→20s
  OnAction: $ACT croaks a bit.
	  : $ACT leaps high into the air.
	  : $ACT hops around a bit.
	  : SAY Ribbit!
	  : N
	  : S
	  : E
	  : W

  It is a small, lime green frog. It's slimy and croaks. What more can you say?
  %%

The zone file documentation has been updated to reflect these changes. Existing
OnAction lists in zone files have been updated to use the new $ACT command.

NOTE: This change will require authors to update their own zone files for the
new behaviour.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>