cmd: If a player is examined make them the participant+notify
This change pulls out the test to see if a Player is being examined so
the result can be used in multiple places. If a Player is being examined
they become the participant and a message is sent to them. For example:

  Diddymus studies you.

We could also use the isPlayer result to avoid sending the message to
the participant when not examining a player, but this needs further
looking into to see the code readability vs performance trade off.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>