cmd: Don't stop SAY processing if Player talking to themself
Currently if a Player talks to themself (no other Players in same
location) processing stops. This means that Players in nearby locations
do not get the 'You hear talking nearby.' message.

This change updates the processing so that nearby Players are notified,
even if a Player is talking to themself.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>