zones: Fix data race by locking destination in checkDoorsHaveOtherSide
In the function checkDoorsHaveOtherSide we have to lock the destination
location otherwise it will cause a data race later on then the
destination location is later read.

NOTE: This is a quick fix as the order of locking the locations is not
done in lowest to highest UID order. If we start loading zones in
parallel this could cause a deadlock.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>