attr,has,cmd: Pass actor to has.Vetoes.Check method
Currently when checking vetoes there is no way to access information
about the actor trying to perform a command. This change allows the
actor to be passed to the Check method allowing additional information
such as location, items carried, etc to be used to determine if a
command should be vetoed or not.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>