attr: Correct freed slices and counters in Inventory.Free
In Inventory.Free all of the content and disabled slice elements are set
to nil to aid garbage collection as the elements are interface
references. However the nil elements should not be accessable still, the
elements should be removed. This change reslices the content and disable
slices to have a length of zero. This change also resets the split and
playerCount values.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>