attr: Fix Player.Check JUNK veto
Since commit 9384781121 (Pass actor to has.Vetoes.Check method) the
signature for Player.Check is longer implementing the has.Vetoes
interface. As a result it is possible to JUNK other players. This change
fixes the signature of Check and adds a compile time check to make sure
we are implementing has.Vetoes.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>