attr/cmd: Fix nil pointer panic in $CLEANUP/$RESET/JUNK commands
At the moment it is possible for $CLEANUP and $RESET events to trigger a
nil pointer panic. This is due to the event holding a reference to the
actor when the event is created. When the event fires the actor may be
stale due to a parent container being cleaned up or reset in the
meantime.

This change now passes the parent Thing of the containing Inventory and
the unique alias of the Thing in the Inventory to be cleaned up or
reset. Care is then take to make sure the actor is not stale before
referencing it. The item to clean up or reset is then searched for
within the actor's inventory. This also ensures that locks are taken
properly on the actor before manipulating the item in question.

On the down side, this makes $RESET and $CLEANUP different to other
commands. It also means there is more locking/relocking to be handled as
the cmd.state is not setup as usual for commands.

This change also fixes the rare situation where a player can be in a
container when an item resets and no message was being seen.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>