zones,docs,data: Add boolean 'Disabled' field to zone header record
This change adds an easy way to disable zones and prevent them from
being loaded when the server starts. It adds a 'Disabled' field to the
zone header record. If set to true, the zone will not be loaded. If the
field is omitted, set to false or left unset the zone will be loaded as
normal.

This can be useful for switching between different versions of a zone
under development or for switching between zone files for special
occasions.

Existing data files have been updated with 'Disabled: FALSE'. While
unnecessary, omitting or not adding the field would have the same
effect, it means the flag can be toggled between TRUE/FALSE easily.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>