attr: Add FindAllVetoes function
When checking for vetoes it is possible that there may be more than one
source for a Thing. For example a location could have vetoes from a Door
and Veto attributes. This change adds a FindAllVetoes function which
returns all attributes implementing has.Vetoes. A separate change needs
to be made to switch existing veto checks to use FindAllVetoes instead
of FindVetoes, at which point FindVetoes can be dropped.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>