attr,has,docs: Add Barrier attribute
This change adds barriers to WolfMUD. Barrier prevent actors from moving
in a specific direction, based on the actor's aliases and the aliases
defined in the barrier's deny and allow lists of aliases.

A single Barrier attribute may be directly added to a location. Multiple
Barrier can be added to a location if they are defined as narratives or
items and put in the location. Items with a Barrier attribute result in
a portable barrier.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>