cmd: Fix get when actor has no Inventory
If an actor has no inventory and tries to get something it will cause a
panic. This is due to an incorrect nil test, which should be using
Found() as we will actually have a nil interface and not a bare nil.

This fixes the issue of mobile using ia scripted GET without having an
inventory defined.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>