attr: Clear, not set, origins on reset (Fixes nil panics)
Currently a reset can sometimes cause a nil panic to be thrown. This is
due to trying to reset at an origin that no longer exists.

When disabled items in inventories were introduced a subtle change in
the mechanics happened. Previously a copy of the item was made and
scheduled for a reset and the original given to the acquiring player or
mobile. However, with disabled items the copy is now given to the
acquiring player or mobile, and the original scheduled for a reset.

With the new mechanics we should not be recalculating the origins for
the copied item. Instead the origins of the copied item should be
cleared. This also lets us reset the original with much simpler logic as
the original does not have to be messed with to correct inventories or
other references.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>