attr: Show Nowhere v no name in Locate Dump where/origin
Currently Locate.Dump shows Nowhere for both a nil or typed nil
where/origin and also when there is no name attribute available.

When debugging using #DUMP it is useful to be able to differentiate
between where/origin pointing to nowhere and there being no name
attribute available - as in the recent case debugging nil panics in
event handling.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>