cmd,doc: Improve PUT/TAKE veto handling + add PUTIN/TAKEOUT pseudo cmds
This change makes sure that all sources of vetoes are checked, not just
the first source found.

The PUT command no longer checks for a DROP veto on the item. The TAKE
command no longer checks for a GET veto on the item.

The PUT/TAKE commands also check for a new PUTIN/TAKEOUT pseudo command
that can be used to create 'one-way' containers.

The PUT command has a new PUTIN pseudo command that can be vetoed. If
PUTIN is vetoed on a container it prevents items from being put into the
container. Specific items can still veto PUT to prevent being put into a
container.

The TAKE command has a new TAKEOUT pseudo command that can be vetoed. If
TAKEOUT is vetoed on a container it prevents items from being taken from
the container. Specific items can still veto TAKE to prevent being taken
from a container.

Documentation in zone-files.txt for VETO/VETOS updated for new PUTIN and
TAKEOUT pseudo commands.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>