cmd: WHICH command with experimental item searching
This commit adds a new WHICH command that uses a new, experimental
search for identifying items. Syntax for using WHICH is:

  WHICH <item, ...>

WHICH will then try to identify all items carried or at the current
location that match. Matches are based on an items alias and additional
qualifiers, which are aliases preceeded with a '+' in the zone files.
Some examples, assuming the following items with aliases and qualifiers,
given in '[]' :

  a small green ball  [+SMALL +GREEN BALL]
  a large green ball  [+LARGE +GREEN BALL]
  a small red ball    [+SMALL +RED BALL]
  a large red ball    [+LARGE +RED BALL]
  a blue box          [+BLUE BOX]
  a red box           [+RED BOX]

Then the following would be matched:

  WHICH BALL
    You see a small green ball here.
    You see a large green ball here.
    You see a small red ball here.
    You see a large red ball here.

  WHICH BOX
    You see a blue box here.
    You see a red box here.

  WHICH GREEN BALL
    You see a small green ball here.
    You see a large green ball here.

  WHICH SMALL BALL
    You see a small green ball here.
    You see a small red ball here.

Multiple items may also be used:

  WHICH BALL BOX
    You see a small green ball here.
    You see a large green ball here.
    You see a small red ball here.
    You see a large red ball here.
    You see a blue box here.
    You see a red box here.

  WHICH SMALL BALL BLUE BOX
    You see a small green ball here.
    You see a small red ball here.
    You see a blue box here.

  WHICH SMALL GREEN BALL BOX
    You see a small green ball here.
    You see a blue box here.
    You see a red box here.

If an item is not identified an attempt will be made to recover:

  WHICH GREEN BOX GREEN BALL
    You see a small green ball here.
    You see a small red ball here.
    You see no 'GREEN BOX' here.'

  WHICH GREEN BALL RED FROG BOX
    You see a small green ball here.
    You see a large green ball here.
    You see a blue box here.
    You see a red box here.
    You see no 'RED FROG' here.'

For testing purposes some new items have been added to the tavern in the
Zinara zone while some have add qualifiers added.

It should be noted that at this time the alias attribute has not been
amended to ignore qualifiers. This means that using 'GET +GREEN' can be
used to get a green ball.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>