cmd: Add count limiting and specific instances to item searching
This commit adds count limiting to item searching. For example, if
'WHICH BALL' returns:

  You see a small green ball here.
  You see a large green ball here.
  You see a small red ball here.
  You see a large red ball here.

Then 'WHICH 2 BALL' would return the first two balls:

  You see a small green ball here.
  You see a large green ball here.

While 'WHICH 1 RED BALL' would return:

  You see a small red ball here.

In addition the first qualifier can be a specific instance such as 1st,
2nd, 3rd, etc. For example 'WHICH 2ND BALL' would return:

  You see a large green ball here.

While 'WHICH 2ND RED BALL' would return:

  You see a large red ball here.

The limit or instance must be the first qualifier for an item.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>