cmd: Add new lastLeadingDigit helper function
Currently SearchLimit uses strings.LastIndexAny to determine if a
qualifier is possibly a special numeric qualifier. However this is not
going to work for ranges of the form n-N, for example: 12-34, as it will
find the 5 and not stop at 2, 5 being the last digit in the string.

This commit adds a lastLeadingDigit function that returns the position
of the last leading digit in a string. For example with 12-34 it will
stop at 2. This also works for numbers with a postfix such as '1st' and
'2nd' where it will stop at the 1 and 2 respectively.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>