cmd: Use min/max limits + add range matches to SearchLimit
This commit switches from using a maximum limit to match or a specific
instance to match to using just a minimum and maximum limit to match.
Especially as a specific instance to match such as '3rd' is equivalent
to the range 3-3.

This commit also introduces processing for ranges of the form n-N. For
example 'WHICH 2-3 BALL' could match the 2nd and 3rd balls.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>