cmd: Move item matching from which.go to matcher.go
This commit moves all of the item matching code into its own matcher.go
file as it can be used by any commands. Also renamed methods MatchXXX
instead of SearchXXX as search is too generic and the code is actually
matching a list of words against items. Yes, it's still searching but we
might want to use search for an actual command later on.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>