cmd: rewrite MatchLimit in matcher
The current MatchLimit function is an unmaintainable mess with
insurmountable issues and failing corner cases. This commit is a
complete rewrite of MatchLimit. It uses a slightly different approach in
that it extracts a set of matches by alias first and then refines that
subset by trying to match qualifiers. Hopefully this version is much
easier to follow and understand.

The MatchXXX functions now return a []Result that allows non-matches to
be sequenced correctly with results. Also allows for non-matches where
there is not enough results for the selected limits.

The specialQualifier function has been updated so that the returned
minimum and maximum values are more suitable for slicing using
slice[min:max]. This reduces the number of calculations that MatchLimit
has to perform on the results.

The WHICH command has been updated for the new functionality.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>