cmd: Fix for 0, 0-0 and 0th limits in matcher
Currently the item search and matcher treats  a limit of 0 or 0th the
same as ALL. However, if you are asking for zero of something then you
should get nothing. This change now returns a NotEnough error in the
results for the limits 0, 0-0 and 0th while allowing zero at the start
of a range 0-n where n > 0. For example 0-1 returns the first match, 0-2
the first two matches etc.

  >which 0 ball
  You don't see that many 'BALL' here.
  >which 0-0 ball
  You don't see that many 'BALL' here.
  >which 0th ball
  You don't see that many 'BALL' here.
  >which 0-1 ball
  You see a small green ball here.
  >which 0-2 ball
  You see a small green ball here.
  You see a large green ball here.
  >

We could perhaps do with a different message for the zero instance but
for now it will have to do :(

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>