attr: Add Things type + *Things.Free method
This commit adds an attr.Things type, defined as a []*Thing. This allows
methods to be defined on Things so that lists of *Thing can be easily
processed. At the moment only a *Things.Free method has been defined
which frees all of the *Thing passed to it.

This addition may be a bad idea, it may also be incorrect. Would it be
better or more useful if Things was []has.Thing? Defining Things as
[]has.Thing would start to blur the lines between the has and attr
packages. Leaving as []*Thing for now, we can change to []has.Thing if
needed at a later date.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>