text,messages: Change to using hex instead of octal notation.
Currently octal notation is used for CSI control sequences. For example
the octal '\033' is used for the control code ESC. However, this can be
painful when debugging as using the '%+q' verb to print out data uses
hex notation. This change therefore updates the sources to use hex
notation so that ESC, octal 033, is now coded as hex 0x1b. To reduce the
amount of hard coded value the constants ESC and CSI have been defined
in colour.go and the foreground and background sequences use these
constants.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>