comms: Clean up error handling in client
This commit cleans up the error handling by using a more succinct swutch
instead of nested if/else. It adds 'connection error:' when reporting
the EOF case. If the error is caused by closing the front end this is
not actually an error so remove 'connection error:' in that case.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>